Add upgrade API
Add Game.addUpgrade function which takes: name of the upgrade "name", description of the upgrade "description", cost of the upgrade "cost", and whether or not the upgrade has been exhausted yet "exhausted" (boolean).
Additionally there are Game.showUpgrade(upgrade) and Game.activateUpgrade(upgrade) which are used to activate the upgrade on a certain condition, and display the upgrade on a certain condition.